+++ karpenkov.pavel/S2/exprProcessing.hpp
@@ -1,23 +1,24 @@
 #ifndef EXPR_PROCESSING_HPP
 #define EXPR_PROCESSING_HPP
 
-#include <iostream>
 #include "queue.hpp"
 #include "stack.hpp"
+#include <iostream>
 
-using queueExpr = karpenkov::Queue< std::string >;
+using queueExpr = karpenkov::Queue<std::string>;
 
 bool isNumber(const std::string &element);
 bool isValidOperator(const std::string &element);
 size_t priority(const std::string &operation);
 
-karpenkov::Stack< queueExpr > inputCLI(std::istream &in);
-karpenkov::Stack< queueExpr > toPostfix(karpenkov::Stack< queueExpr > expressions);
+karpenkov::Stack<queueExpr> inputCLI(std::istream &in);
+karpenkov::Stack<queueExpr> toPostfix(karpenkov::Stack<queueExpr> expressions);
 
-karpenkov::Stack< int > calculateExpr(karpenkov::Stack< queueExpr > &postfixExpr);
+karpenkov::Stack<long long>
+calculateExpr(karpenkov::Stack<queueExpr> &postfixExpr);
 
-void queueOutput(std::ostream &out, karpenkov::Stack< int > result);
+void queueOutput(std::ostream &out, karpenkov::Stack<int> result);
 
-int eval(int a, int b, const std::string &operation);
+long long eval(long long a, long long b, const std::string &operation);
 
 #endif
